exports to default exports. setupGenCode for new file. Types exports moved to files
