+++ dev/src/SmartC/shaper/templates.ts
@@ -1,12 +1,12 @@
 import { SC_FUNCTION } from '../typings/contractTypes'
 import { REGISTER_TYPE_DEFINITION, LONG_TYPE_DEFINITION, STRUCT_TYPE_DEFINITION, MEMORY_BASE_TYPES, MEMORY_SLOT } from '../typings/syntaxTypes'
 
-type ObjectType<T> = T extends 'register' ? REGISTER_TYPE_DEFINITION :
+type ObjectTypeDefinition<T> = T extends 'register' ? REGISTER_TYPE_DEFINITION :
     T extends 'long' ? LONG_TYPE_DEFINITION :
     T extends 'struct' ? STRUCT_TYPE_DEFINITION :
     never;
 
-export function getTypeDefinitionTemplate<T extends 'register'|'long'|'struct'> (templateType: T) {
+export function getTypeDefinitionTemplate<T extends 'register'|'long'|'struct'> (templateType: T) : ObjectTypeDefinition<T> {
     let retObj: REGISTER_TYPE_DEFINITION | LONG_TYPE_DEFINITION | STRUCT_TYPE_DEFINITION
     switch (templateType) {
     case 'register':
@@ -42,7 +42,7 @@ export function getTypeDefinitionTemplate<T extends 'register'|'long'|'struct'> 
     default:
         throw new TypeError('Internal error')
     }
-    return retObj as ObjectType<T>
+    return retObj as ObjectTypeDefinition<T>
 }
 
 export function getMemoryTemplate (memType: MEMORY_BASE_TYPES) : MEMORY_SLOT {
