+++ dev/src/SmartC/smartc.ts
@@ -1,10 +1,10 @@
-import { preprocess } from './preprocessor/preprocessor'
-import { tokenize } from './tokenizer/tokenizer'
-import { parse } from './parser/parser'
-import { shape } from './shaper/shaper'
-import { syntaxProcess } from './syntaxProcessor/syntaxProcessor'
-import { codeGenerate } from './codeGenerator/codeGenerator'
-import { byteCode } from './byteCoder/byteCoder'
+import preprocess from './preprocessor/preprocessor'
+import tokenize from './tokenizer/tokenizer'
+import parse from './parser/parser'
+import shape from './shaper/shaper'
+import syntaxProcess from './syntaxProcessor/syntaxProcessor'
+import codeGenerate from './codeGenerator/codeGenerator'
+import byteCode from './byteCoder/byteCoder'
 
 import { PRE_TOKEN, TOKEN } from './typings/syntaxTypes'
 import { CONTRACT, MACHINE_OBJECT } from './typings/contractTypes'
